Refactor import paths for chat and dashboard pages
